backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
groq==0.13.1
mistralai==1.4.0
pymupdf==1.25.3
sentence-transformers==3.4.1
python-multipart==0.0.20
httpx==0.28.1
